feat: add jwt secret for authentication (#243)
* feat: Add jwt secret for authentication

The code changes add a new field `jwt_secret` to the `Cli` struct in `lib.rs`. This field allows setting the JWT secret for authentication. Additionally, the `create_router` function in `api/mod.rs` now takes the `jwt_secret` as a parameter.

Recent user commits and repository commits are not relevant for generating the commit message.

* feat: jwt as option without default value

* feat: add dummy token

* fix: lock file permission